requirements.txt
streamlit==1.38.0
phonenumbers==8.13.47
pypdf==5.0.0
Pillow==10.4.0
opencv-python==4.10.0.84
pydub==0.25.1
faster-whisper==1.0.2
imageio==2.35.1
imageio-ffmpeg==0.4.9
